KEYCLOAK-XXX Add support for merging user accounts
We now allow users to merge their currently logged in user
account with another user account which they own.

A user can go to the "Merge Account" page in the account
application and enter the credentials of an account that
should be merged into the currently logged in account.

If an account for the given credentials was found the
account settings of this account are merged into the current
account. Merged user information includes assigned client- and
realm-roles as well as user profile information (if not already
present in the currently logged in account).