Explicit types for ValidationKey and ValidationContextKey instead of Strings
We now use an interfaces to denote validation keys and validation context keys.
For ease of use an enum is used to define default values.
API users can reuse the provided definitions or create their own
by defining an enum that implements the ValidationKey
or the ValidationContextKey interface. It is recommended to use a singleton
for a *Key implementation, hence an enum is a good choice.